Move qpid scripts ({gutterball->server}/bin/qpid)
I also removed gutterball references from the script.

With this change, adding `-q` as an argument to `server/bin/deploy` will
cause qpid to be automatically set up. Furthermore, if both `-q` and
`-f` are provided, it will regenerate QPID certs as well as other certs.

This should be followed up by an update to candlepinproject.org docs.