Annotate User object as a "split" model.
Because the User object is meant to take a password in but never send a
password out, the versions of the User object used during serialization
and deserialization are asymmetric.  The Swagger client generation
library does not handle this situation very gracefully.  The recommended
solution is described in
https://github.com/swagger-api/swagger-core/issues/1532 and
https://github.com/swagger-api/swagger-core/issues/1590

Following the recommendation, this patch creates a dummy object,
UserCreationRequest, that we tell Swagger to use as the parameter for
the POST /users resource call.