Add and correct Swagger annotations so Swagger code-gen works better.
* Since the CandlepinQuery<T> return type is resolved into an array of
type T via a JAXRS interceptor, we need to explicitly inform Swagger
that the return type is actually a list of Ts.

* Provide objects submitted via an HTTP request's body with a name and
mark them as required if appropriate.  By giving a name, the generated
client method will expect a parameter instead of just taken a more
generic options hash (in the Ruby client at least).