1441700: Fix issues with Status caching invalidation
Sometimes the JCache we used for caching the Status
would not invalidate after the configured 5 seconds.

Because of this, we implemented our own caching
mechinism for Status due to cache invalidation
issues we were having with the ehcache-jcache
implementation.

The new StatusCache uses a static Status instance to
track the latest computed candlepin status.
When a status is set in the cache it has a
5 second lifespan.