Database failfast
This commit extends the current failfast check that runs every 10 secs
and uses it to check if the connection to database is ok.
If it is not, SUSPEND mode is entered and vice versa.
Before the transition from SUSPEND to NORMAL mode due to DB recovery
all database connection in the pool will be closed.