Add spec test for prioritization of host based pools
This adds a new spec test for the following case.

There exists a Guest A on Host A. Guest A has two products
installed (let's call them 1 and 2). The Org that Guest A and Host A
are in has a pool available that would provide both products 1 and 2
if used to entitle Guest A. Host A has two entitlements that create
pools of unlimited quantity each providing just one of the required
products.

An auto attach request is made by Guest A.

The test fails if the auto attach request results in attachment of
anything aside from one entitlement from each of the pools provided
by Host A.