Allow guests in one org to look for hosts in another org
If a hypervisor is sharing its host-restricted guest pools, guests in
the recipient organization will need to be able to look for the host in
the originating organization.  This patch adds that capability, but it
is likely to perform poorly given that it iterates through all the
recipient's pools in order to deduce the originating organizations.