Search for an guest's host across all orgs in most cases.
Care must be taken not to expose host organizations to guests that do
not actually have permission to view them.  We need to search across all
orgs because organizations do not have an easy way to know what
organizations are sharing with them.  This knowledge could be
provided via a join table but after discussion with the team, we decided
those additional relationships created complexity that was unnecessary
at this time.