Awood/multi org hypervisors (#1594)
* Allow guests in one org to look for hosts in another org

If a hypervisor is sharing its host-restricted guest pools, guests in
the recipient organization will need to be able to look for the host in
the originating organization.  This patch adds that capability, but it
is likely to perform poorly given that it iterates through all the
recipient's pools in order to deduce the originating organizations.

* Search for an guest's host across all orgs in most cases.

Care must be taken not to expose host organizations to guests that do
not actually have permission to view them.  We need to search across all
orgs because organizations do not have an easy way to know what
organizations are sharing with them.  This knowledge could be
provided via a join table but after discussion with the team, we decided
those additional relationships created complexity that was unnecessary
at this time.

* Test that guests in one org don't autoheal hosts in another org

Keeping guests from provoking an autoheal when subscriptions are shared
across organizations is actually handled both in Entitler.bindByProducts
and in CandlepinPoolManager.entitleByProductsForHost.  In both places,
having a guest and host in different organizations is sufficient to
abort the healing process.