1457902: Fix failure listing Jobs when class not found
When listing jobs, if the job class was deleted and is not
found, hibernate would through an exception when building
up the JobStatus object.

This was addressed by changing the JobStatus.jobClass field
to be a String instead of a class. This way hibernate will
not try and create a Class object from the class string.

Callers of JobStatus.getJobClass will now have to instantiate
the Class directly and handle cases where the class is not found.