Add buildr plugin to download swagger.json and generate client
This plugin defines several commands:

* swagger:json - This command downloads swagger.json from the location
specified by swagger.json_source in the buildfile.  The swagger.json
file is placed in the top level of the project.  Candlepin **must** be
running for the download to be successful.

* swagger:client - This command attempts to generate a REST client using
swagger-codegen and a swagger.json file located in the top level of the
project.  Several options can be specified such as 'lang' to define the
language to use for the generated client and 'destination' to specify
where to write the resulting client files.

* swagger - This command simply combines the previous two commands for
convenience.