Fix custom pom plugin to be compatible with new Buildr version.
The new version of Buildr added attributes to its POM plugin that we did
not have in our POM plugin.  When the IDEA plugin attempted to access
these attributes, an exception was thrown.  This patch adds in those
attributes and writes them, if defined, into the POM file the same way
that Buildr would itself.