1476897: Choose host_specific/virt_only pools first
This is a change to the rules that allows virtualized systems to
reliably receive entitlements from host_specific and/or virt_only
pools before the consumption of quantity from physical pools.

There is one case that this change could make a bit worse.
In the case that there are not enough host_specific and/or
virt-only pools available to cover the products on the guest
system, the guest will be given entitlements from all available
virtual pools that provide at least one product in addition to
receiving an entitlement from as many physical pools as necessary
to cover the products installed.

If the available physical pools included one that would provide
all the installed products on the guest, all the entitlements
from the virtual pools would be redundant.

Adds a spec test for host pool prioritization