Fix custom pool lookup for bonus pool
the absence of source subscription is not a reliable way to determine if a pool is a custom pool,
since that behaviour varies with the candlepin verion. More reliable is the presence or absence of
an upstream pool id.

CandlepinPoolManager.isManaged is now the single method to determine if a pool is custom, and checks
both source subscription and upstream pool id.