1497288: Mark jobs canceled when shutting down
Jobs are tracked in both cp_job and in the Quartz tables.  Prior to this
patch, when Pinsetter shutdown, it deleted job detail information from
the Quartz tables but did not alter the job in cp_job.  Upon startup,
the UnpauseJob would attempt to restart the job found in cp_job but no
correlating record would be found in the Quartz tables.

This patch adds a line to mark a job as canceled in the cp_job table as
the job is purged from the Quartz tables.  It also invokes a JobCurator
method that removes orphaned jobs on start up although this measure is
primarily preventative.