Add a dummy assembly.xml to appease maven.
Our buildr pom plugin configures every sub-project to use the
maven-assembly plugin and to expect an assembly.xml.  We use this
configuration because Candlepin has a special definition for a source
artifact.  The "checks" subproject doesn't need a special assembly since
all we want is the jar file, but configuring the buildr plugin to
discriminate which project should have the assembly plugin added to the
pom and which shouldn't is more complicated than just adding an assembly
file for the checks subproject.

The assembly file added is just a copy of Maven's "src" assembly as
defined at
http://maven.apache.org/plugins/maven-assembly-plugin/descriptor-refs.html#src