Event Simplification: Removed oldEntity from Events
- Removed the oldEntity field from the Event class, also
removed setOldEntity/getOldEntity from EventBuilder and EventFactory.
- Made setEventData on EventBuilder public to be used instead
of setOldEntity when building Events. This method will probably be
re-used similarly when newEntity will be removed in the future and
will be used when creating CREATED/DELETED events. It will be
overloaded for creating MODIFIED events (which currently need both
old/new entity).
- Removed a couple of unused methods.
- Now that events no longer serialize whole Objects for the purpose of storing
the oldEntity field (which, in the case of Consumer, was forcing initialization of lazy-loaded collections as a side-effect),
we need to force their initialization before we get to save HypervisorUpdateJob's result,
to avoid LazyInitializationExceptions when we get to fetch it from the db later.