1514515: Correctly update distributor's modified product content
When a regular consumer binds to an entitlement providing product
modifiers, candlepin matches on product/provided products to
determine if the modified content should be added to the
entitlement certificate.

However, when a distributor binds to the same entitlement, candlepin
will also try to match on the derived product and derived
provided products.

This patch addresses a bug where candlepin was not considering
derived products when checking to see which modifier entitlements
to regenerate when pools/entitlements are deleted.