1517848: Set the KeyPair object to lazy load.
Joining on the cp_key_pair table can take hundreds of milliseconds in
tables with many rows.  The KeyPair object is not even used in most
cases, so switching it to lazy loading will shave some time off our
queries at the expense of an additional query in the cases where we
actually need the KeyPair.