1531627: Reduce possibility of lock waits on Async Hypervisor Check in
This work seeks to reduce the time taken to perform Async Hypervisor Checkins
and to reduce the scope of the locks acquired for a given hypervisor checkin.

Previously we had seen lock wait timeout's occurring on Consumer Type between
otherwise unrelated (meaning different data, different Org etc) async hypervisor
checkins. This appeared to be caused by the inclusion of Consumer Type in the
getHypervisorsBulk query (wherein a select for update was used to lock the rows
retrieved).

Now HypervisorUpdateJobs flush queries to the database once per job.
Consumer compliance status is calculated only once per consumer (was twice before).
The owner is no longer queried from the database once per hypervisor consumer.
HypervisorUpdateJobs no longer store a serialized version of the entire consumer
in the JobStatus (as we really need only the list of consumer uuids for each state
Created, updated, etc).
The GuestIds that have been migrated from one consumer to another are updated
directly.

The remainder of the work that has gone into this fix was to speed up the
checkins. Further commit messages of this work are below:

* Adds lazy fetch of consumer type to getHypervisorsBulk

* Adds new HypervisorUpdateResultUuids class (no need to serialize
  whole consumers)

* No longer call consumerResource.create/performConsumerUpdates

* Use ownerId when looking up hypervisor consumers by owner