Add a unit test to ensure all changelogs are version 3.1
Using a schema version more recent than what's packaged in the Liquibase
JAR file results in an attempt to download the missing XSD from the
internet.  For machines that are not connected to the internet, a fatal
error results.  This test is meant to prevent invalid schema versions
from creeping into our code.