Qpid reconnect attempts now have a constant interval.
Previously, the class responsible for attempting to reconnect to Qpid
would begin with a small interval between attempts and then gradually
grow that interval so that reconnect attempts would grow farther and
farther apart.

This commit removes that functionality and replaces it with a constant
interval between reconnection attempts.  The previous algorithm provided
little tangible benefit and added additional code and configuration
complexity.