1546236: Properly handle connections when they are recreated
Whenever the QpidConnection needed to be rebuilt, the existing
connection was being disregarded and a new one created. The old
connections were not getting cleaned up causing candlepin to
exhaust all connections available to Qpid.

This was primarily surfacing when HornetQ would try to resend
failed messages but the Qpid queue was in FLOW_STOPPED. In
this case, QpidConnection would be marked as stale because of the
exception, and would create a new connection to Qpid, leaving
the other around. In the case of FLOW_STOPPED, a new connection
isn't required - we just need to wait for the queue to catch up.

Since the exceptions thrown by Qpid for the various failure
cases, we could never tell exactly when to recreate the connection
or when the connection was Ok.

This patch addresses this PR by monitoring the Qpid status and
only creating new connections when required as well as cleaning
up after itself.

Notes:
- Pulled Qpid connection checking via QMF into an isolated service
  separate from SuspendModeTransitioner.
- Added a listener framework for objects to be notified when the latest
  Qpid status is determined.
- SuspendModeTransitioner now listens for Qpid status updates instead
  of running the checks itself.
- QpidConnection now listens for status updates so that it can better
  manage its connection state (connect, reconnect)
  * When Qpid is FLOW_STOPPPED, block message sending to Qpid until
    it is notified that Qpid is able to accept the message.
  * When notified that Qpid went down, close the old connection so that
    it is no longer hanging around.
  * When notified that Qpid is back up, re-create the connection again.