Add toString implementations for objects stored as job results.
These classes are the results of various jobs and currently they are
stored in the cp_job result column by running toString on them.  The
result column is meant to be a quick way to check a job's outcome while
the resultData column holds the entire body of the results.  Using the
default toString implementation defeats the purpose of the result
column.