Refactor PKIReader to decouple from BouncyCastle.
Previously we were delegating all of the work required to read a private
key out of a PEM file to BouncyCastle.  To loosen our coupling to
BouncyCastle, this commit performs a lot of the required work itself.
This commit reads the PEM files and dispatches the decoding and/or
decryption of the key to a class based on the format of the key.
Currently, the PKCS1 classes still use BouncyCastle to parse the ASN1
and to decrypt the non-standard OpenSSL encrypted PKCS1 format.