Modify ResultDataUserType to read legacy format or JSON and write JSON
We want to move away from storing binary serialized objects in the
database, but in a backwards-compatible way.  ResultDataUserType will
now always write Jackson-generated JSON into the resultData column of
cp_job and read JSON data from the column.  However, if there is an
error in reading the JSON, it will fall back to attempting to
deserialize the bytes as a Java object.

ResultDataUserType is also configured to deserialize the JSON into the
correct Java type based on how the field is declared in the entity
class.  If Jackson has no mapping to that type or encounters another
issue, it will deserialize as best it can (generating a Map of the
key-value pairs for example) and return the result as an Object.