Insert owner level logging filter only when needed.
The LoggerAndMDCFilter is responsible for logging messages at a given
level for specific owners.  We use this tool in production to get DEBUG
level messages for specific owners without having to log everyone else's
requests.  The filter is fast: the decide() method takes 0.1ms or less,
but it is still being run for *every* logging call.  Those tenths of a
millisecond add up over millions of invocations and all for a feature
that is used only rarely.

This patch inserts the LoggerAndMDCFilter when owner level logging is
requested and removes it when owner level logging is disabled.  It also
checks at servlet start-up time to see if owner level logging is
enabled anywhere and adds the filter if so.