Remove toString call where it is not necessary.
The beauty of SLF4J style interpolation is that if the log level is
higher that the method requires, it will not call toString on the object
and therefore saves some time.  By calling toString on the object before
it is passed in as a parameter, we lose those gains.