Maintain order of messages sent to Qpid
When Qpid is down or in flow_stopped, messages that are
received from Artemis are rolled back and enter the
resend on failure loop. When this happens, messages have
the potential to be sent to Qpid out of order due to retry
delays.

This patch shuts down the ClientConsumer that receives the
messages as soon as it is notified that Qpid is in a bad
state. When Qpid comes back up, the session is restarted
and all messages are sent to Qpid in order.

All EventListeners are now required to implement a method
defining whether Qpid is required. If it does, it will be
notified whenever the Qpid connection goes down, or Qpid
is in a bad state.

There are now two MessageHandler implementations that are used
to determine how to handle messages that they receive.

1) EventMessageReceiver is the default message handler. On failure
it will put the message back on the queue and will retry delivery
on a configurable interval.

2) QpidEventMessageReceiver is used for messages that are intended
to be forwarded on to Qpid. If a message does not make it to Qpid
due to a connection issue, the message is put back at the start
of the queue and will be redelivered in the order it was added once
the ClientConsumer is restarted, or candlepin is restarted. If a
message send fails due to any other exception, the message is logged
and discarded.

In order for the retry logic to behave in the way we intend, the
QpidEventMessageHandler needs to on an address that has custom
address settings. In order to acheive this without having to have
the producer send multiple messages in a loop, messages send to the
event.default address will get forwarded to event.qpid via use of
a non-exclusive diverter. Diverted messages will be received with
a different message ID, however, candlepin will log the original
ID along with the new when the message is received.

This PR also merges the ListenerWrapper into the EventReceiver
as the wrapper was not required.