Move BouncyCastle specific methods into self-contained class
PKIUtility is now a concrete class that implements crypto functions only
using the JRE's JCA API.  For more complex functions that require access
to the underlying crypto library it delegates to an implementation of
PKIProviderUtility.

This commit also removes some uses of the Jukito test runner.  For one,
we didn't need it in the first place, and for another, it was breaking
after I converted PKIUtility to a concrete class because it was trying
to inject all the private fields marked with @Inject before our tests
had a chance to bind CertificateReader to the stub implementation we use
in testing.