Add JSS as a cryptographic provider.
This commit adds a JSS implementation of most of our cryptography
classes that coexists alongside the BouncyCastle implementations we
already have.

It does not handle the streaming CRL process.  That will be in future
commits.  As well, the JSS implementations are not used since Guice does
not bind them as the implementation for the PKIProviderUtility class.

This commit also:

* Requires JSS as an RPM dependency.
* Changes Key to RSAPrivateKey in most cases. Candlepin only works with
  RSA keys.  Pretending otherwise is pointless.  This patch changes most
  of the method signatures to deal with RSAPrivateKey objects and handle
  casts from PrivateKey objects (returned from the KeyPair class) to
  RSAPrivateKey objects.