Separate crypto provider specifics in X509CRLEntryStream.
Most of X509CRLEntryStream is not provider specific but since the JDK
defines X509CRLEntry as an abstract class we need to create an anonymous
implementation based on BouncyCastle's CRLEntry.

This patch also moves toward using crypto classes that are present in
the JDK instead of provider specific ones.  We were using BouncyCastle's
ContentSigner instead of the JDK's Signature.  Likewise with CRLEntry
instead of X509CRLEntry.