Have Checkstyle do differential checks based on suppression.xml
The eclipse-checkstyle plugin had a byzantine way of configure checks to
run differently on different groups of the code (e.g. "main" versus
"test").  The plugin called for defining properties in a single check
and then resolving those properties.  To replicate the effect in Buildr,
we would pre-process the XML file, resolving the properties with Ant.

This elaborate process makes life much harder than it has to be.  By
using a suppression file, we can tell Checkstyle not to run certain
checks on files matching certain criteria.  Not only is this easier, but
it has the benefit of working with IntelliJ's Checkstyle plugin.

This commit adds a suppression.xml and sets Checkstyle to use it.  Much
of the cruft around the eclipse-checkstyle plugin remains however as our
Buildr plugin is tightly coupled to the idioms of eclipse-checkstyle.
Cleaning all that up is left to another day.