1584256: Detach consumers from persistence context to avoid OOM
Updating the status of a large number of consumers (roughly 10000) could
run the JVM out of memory due to all the consumers referenced in the
consumerStackedEnts map.  During status calculation, the consumer facts
are accessed and during those accesses, the Hibernate proxy object is
replaced by the actual map of facts.  The consumerStackedEnts map would
accordingly grow as each status was calculated.

This patch detaches several model objects as soon as we stop needing
them and most importantly, it detaches all the consumers in the
consumerStackedEnts map.  The consumers are then fetched again, receive
an updated status, and are then immediately detached to keep the
persistence context at roughly a constant size.