Silence all hibernate deprecation warnings
After a hibernate upgrade, the log file is not getting
polluted with deprecation warnings due to our use of
the old Criteria API. This patch disables those warnings
so that the log file doesn't grow when deployed in production.

The warnings have also been silenced when running unit tests
as well so that the log file isn't polluted.