Add Immutable Entity & Hibernate 2nd level cache support
Remove Clear from pool providedProduct & DerivedProvided Product setters as it can causes extraneous selects to the database.

If we are setting values in a pool object but are never going to persist them pre-clearing the provided & derivedProvidedProducts causes extraneous selects. This was done in ProductCurator.hydratePoolProvidedProducts() for serialization.

ProductCache has been replaced by usage of the Hibernate L2 Cache.

Remove old code that modified the dev SKU MCT3295 in a way that is already set upstream.

Limit default cache to 100k entries

Add locking of the OwnerProduct on product updates in order to prevent collisions throwing away changes.

Remove an unnecessary hashset.

Replace * import with individual class imports for OwnerProductResource

Avoid the use of Pool.clear() as it will force an immediate select from the database.