ENT-930 Fix role & addons mismatch error in rules
Fixes a few things on the back of previous work
to make roles & addons behave similarly to products
in terms of compliance:

- Alter rules' use of product attribute name from 'role' to 'roles'
as agreed with IT.
- Check for role/addon compliance against each separate value in the
CSV of the role/addon product attribute strings.
- Added proper Compliance Reason output strings for role/addon
mismatch to show to the user.
- Added spec tests that check mismatched role/addons during direct attach
does not result to failure, the pool is successfully attached,
and the proper reasons are returned for partial match.
- Added unit tests that check mismatched role/addons during auto attach
results in pools not being selected.