Remove emojis from unicode test.
MySQL/MariaDB is terrible and the default UTF8 locale only supports
three byte wide characters.  Thus, emoji support requires special
settings to the database.  An org with emojis in the name is not very
realistic either, so I'm changing this test to use characters that
MySQL's anemic UTF8 locale will support and that will be more true to
reality,