1620226: Perform CRL generation without exceeding memory and maxing CPU
Using a HashSet instead of an ArrayList to store the serials improves
the performance of .contains when checking if an entry in the CRL file
should be removed due to expiry.

Batch lookup/process fetched serials to keep the memory footprint down
and to allow garbage collection of the serials earlier. The number of
records per batch can be configured via the candlepin config file.

NOTE: Due to the way the CRL generation framework works, when batching
the serial processing, we end up re-scanning the crl num_serials/batch_size times.
This isn't ideal, but we will make due until we make the move to generating
a plain text serial file.