Fix pom generation to fix builds
I had to modify the pom generation to include more dependencies in
specific scopes. See the changes to `build.gradle` files for details.

I removed the standlone candlepin-common RPM, and modified its project
to share version with the server specfile.

This also regenerates the pom files to be up-to-date with gradle files.