Send the message as raw bytes & add properties for routing.
The artemis native clients write UTF strings in the format of a 4 byte integer of the length of the body followed by a 2 byte short Integer of the length of the body followed by the actual body. This code is found in the artemis-commons/src/main/java/org/apache/activemq/artemis/utils/UTF8Util.java. This causes understandable concerns for non-native clients or clients that are not using UTF8Util to decode the strings. Performing the conversion in candlepin & sending raw bytes allows us to better support clients other than candlepin that want to consume the messages.