Add supported_architecture fact
The 'supported_architectures' fact is used to provide the information which
architectures (like i386, amd64) a host supports. This is used e.g. on
Debian / Ubuntu hosts to support multiple architectures. A common scenario is to
use amd64 and i386 on Debian / Ubuntu. Therefore, the host needs to be able to install
packages from amd64 and i386 repositories.