Convert request_scanner.rb to Python
Create a Python version of request_scanner.rb to perform the same task
in a language that will still be used in the future.
The implementation mostly follows the Ruby version for the logic, doing
things in a Pythonic way:
- argparse instead of OptionParser
- a custom argparse.Action for getting regular expression objects
  instead of strings
- Python's logging instead of homebrew functions; logging works fine for
  our case with only few config tweaks
- handle errors directly on file opening, and item removal from dict,
  rather than checking beforehand
- pre-compile one more regex that was inline in a condition
- iterate directly on keys of dictionaries in case only the keys are
  needed
- format the code using the 'black' tool, so there is a standard
  codestyle matching the modern Python practices

The resulting output is almost the same; the only difference is due to
the logging and applies to warning lines: instead of "WARN:" there is
"WARNING:".

Side effect of the conversion: the Python version seems faster than the
Ruby version.