Prevent layer 2 cache eviction & improve caching of branding
* Set the query Hint on native queries in order to prevent hibernate from
unnecessarily clearing caches of immutable objects. If a hint is not set
then all cache regions are cleared during native query execution.

* Set the appropriate cache hints on Branding objects so that they are
cached as immutable.

* Update the cache definition to standard 24 hours instead of using seconds
in order to make it more clear for humans reading the file in the future.

* Move the flush from the curators to the managers as the NATIVE_SPACES query
hint changes the default hibernate behavior to no longer do a global flush
before execution.