CANDLEPIN-847: Fixed intermittent failure in RefreshPoolsSpecTest
- Adding sleep should fix the random crash test because
it seemed that the certificate regeneration is working. However,
the operations are too quick in succession, causing both certificates
to have the same timestamp. I also swapped the comparison of the
certificate content and the timestamp in case this test crashes
again, so we will know whether it was due to the timestamp or the
certificates being the same.`