CANDLEPIN-944: Droped FK for cp2_owner_content/product on cp_owner
- When the satellite tried to update from version 6.15 to 6.16, we
decided not to use cp2_ownerproduct/content anymore and to utilize
new tables for that purpose. However, we chose to keep the original
tables in case we needed to revert to the previous version. We didn’t
delete the foreign keys (FK) on those tables, which prevented the
deletion of custom owners due to constraint violations.